pyproject.toml
[project]
name = "qnap-telegram-bot"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "jiz4oh", email = "me@jiz4oh.com"},
]
dependencies = [
    "python-telegram-bot>=21.3",
    "requests>=2.32.3",
    "python-dotenv>=1.0.1",
]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
